카테고리 리스트 및 검색결과 리스트, 주변검색 리스트 내 상세정보 100자 이내 노출 처리 추가 gps 좌표 확인 및 gps 좌표를 통한 반경 5km내 관광정보 검색 기능 개발

+++ app/src/main/java/com/eye2web/travel/DetailInfoActivity.java
@@ -42,8 +42,8 @@ public class DetailInfoActivity extends BaseActivity {
         String contentId = "";
         String contentTypeId = "";
         String areaCode = "";
-        float mapx = 0;
-        float mapy = 0;
+        double mapx = 0;
+        double mapy = 0;
 
         contentId = item.getContentid();
         contentTypeId = item.getContenttypeid();
@@ -64,7 +64,7 @@ public class DetailInfoActivity extends BaseActivity {
      * @Author : Andrew Kim
      * @Description : 상세정보 조회
     **/
-    public void getContent(String contentId, String contentTypeId, String areaCode, float mapx, float mapy) {
+    public void getContent(String contentId, String contentTypeId, String areaCode, double mapx, double mapy) {
         detailApiService = new DetailApiService();
         DetailIntroItem detailIntroItem = new DetailIntroItem();
         DetailCommonItem detailCommonItem = new DetailCommonItem();
